(GH-5) Add creation of NuGet package into build
This is based on who the NuGet Package is created for the Chocolatey
Agent.  UppercuT files have been updated to be the same as that project
as well as removing files that aren't required.

In line with changes to other projects, switch to using the shared
chocolatey.snk file, rather than one specific to the project.